# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Utiva Growth AI Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List all AI workflow automations
      responses:
        "200":
          description: List of workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create a new workflow automation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowBody"
      responses:
        "201":
          description: Workflow created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{id}/toggle:
    post:
      operationId: toggleWorkflow
      tags: [workflows]
      summary: Enable or disable a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{id}/run:
    post:
      operationId: runWorkflow
      tags: [workflows]
      summary: Manually trigger an AI workflow run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workflow run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunResult"
...
components:
  schemas:
...
    Workflow:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [lead_followup, enrollment_reminder, re_engagement, content_push, lead_scoring, drop_alert]
        isActive:
          type: boolean
        triggerCondition:
          type: string
        actionTemplate:
          type: string
        runsTotal:
          type: integer
        runsThisWeek:
          type: integer
        successRate:
          type: number
        lastRunAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - type
        - isActive
        - triggerCondition
        - actionTemplate
        - runsTotal
        - runsThisWeek
        - successRate
        - createdAt
    CreateWorkflowBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [lead_followup, enrollment_reminder, re_engagement, content_push, lead_scoring, drop_alert]
        triggerCondition:
          type: string
        actionTemplate:
          type: string
      required:
        - name
        - description
        - type
        - triggerCondition
        - actionTemplate
    WorkflowRunResult:
      type: object
      properties:
        workflowId:
          type: integer
        affectedCount:
          type: integer
        status:
          type: string
        message:
          type: string
        aiGeneratedMessage:
          type: string
          description: Sample AI-generated message produced by the workflow
      required:
        - workflowId
        - affectedCount
        - status
        - message
...