# lib/api-zod/src/generated/types/workflow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Utiva Growth AI Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkflowType } from "./workflowType";

export interface Workflow {
  id: number;
  name: string;
  description: string;
  type: WorkflowType;
  isActive: boolean;
  triggerCondition: string;
  actionTemplate: string;
  runsTotal: number;
  runsThisWeek: number;
  successRate: number;
  lastRunAt?: Date;
  createdAt: Date;
}